method channel for accessing user info
