+++ auth0_flutter/darwin/Classes/CredentialsManager/CredentialsManagerUserInfoMethodHandler.swift
@@ -1,0 +1,16 @@
+
+import Auth0
+
+#if os(iOS)
+import Flutter
+#else
+import FlutterMacOS
+#endif
+
+struct CredentialsManagerUserInfoMethodHandler: MethodHandler {
+    let credentialsManager: CredentialsManager
+
+    func handle(with arguments: [String: Any], callback: @escaping FlutterResult) {
+        callback(credentialsManager.user)
+    }
+}
