+++ auth0_flutter_platform_interface/lib/auth0_flutter_platform_interface.dart
@@ -25,6 +25,7 @@ export 'src/credentials-manager/options/has_valid_credentials_options.dart';
 export 'src/credentials-manager/options/local_authentication.dart';
 export 'src/credentials-manager/options/save_credentials_options.dart';
 export 'src/credentials.dart';
+export 'src/user_info_credentials.dart';
 export 'src/database_user.dart';
 export 'src/id_token_validation_config.dart';
 export 'src/login_options.dart';
