+++ auth0_flutter_platform_interface/lib/src/credentials-manager/credentials_manager_platform.dart
@@ -1,10 +1,7 @@
 // coverage:ignore-file
 import 'package:plugin_platform_interface/plugin_platform_interface.dart';
 
-import '../../auth0_flutter_platform_interface.dart';
-import '../user_info.dart';
-import '../credentials.dart';
-import '../request/request.dart';
+import 'package:auth0_flutter_platform_interface/auth0_flutter_platform_interface.dart';
 import 'method_channel_credentials_manager.dart';
 import 'options/get_credentials_options.dart';
 import 'options/has_valid_credentials_options.dart';
